[DSpace#8356][DURACOM-69] Change Requests part. 3
