+++ dspace/config/modules/assetstore.cfg
@@ -37,8 +37,7 @@ assetstore.s3.subfolder =
 
 # please don't use root credentials in production but rely on the aws credentials default
 # discovery mechanism to configure them (ENV VAR, EC2 Iam role, etc.)
-# The preferred approach for security reason is to use the IAM user credentials, but isn't always
-possible.
+# The preferred approach for security reason is to use the IAM user credentials, but isn't always possible.
 # More information about credentials here: https://docs.aws.amazon.com/sdk-for-java/v1/developer-guide/credentials.html
 # More information about IAM usage here: https://docs.aws.amazon.com/sdk-for-java/v1/developer-guide/java-dg-roles.html
 assetstore.s3.awsAccessKey =
